docwriter/requirements.txt
attrs==22.2.0
btrfsutil==6.1.2
cffi==1.15.1
cryptography==38.0.4
lxml==4.9.2
packaging==21.3
ply==3.11
pycparser==2.21
PyGObject==3.42.2
pyparsing==3.0.9
python-docx==0.8.11
TBB==0.2
team==1.0
